+++ src/main/java/ca/sait/vezorla/service/AuthenticationServices.java
@@ -4,7 +4,9 @@ import ca.sait.vezorla.model.Account;
 
 public interface AuthenticationServices {
 
-	   public void forgotPassword(String email);
-	    public Account login(String email, String password);
-	    public void logout();
+    void forgotPassword(String email);
+
+    Account login(String email, String password);
+
+    public void logout();
 }
